Replace old VFORK_SUBSHELL define and fix Clang warnings.
I'm going to use PREFER_POSIX_SPAWN instead of VFORK_SUBSHELL for
my future vfork() / posix_spawn() optimization work, so I updated
the config header accordingly.

I also fixed all of the C++20 compiler warnings in locale.cc.

+++ include/shmbutil.hh
@@ -115,7 +115,7 @@ MB_INVALIDCH (size_t size)
 // The necessary state variables are passed by the caller.
 static inline void
 advance_char (std::string::const_iterator &it,
-              const std::string::const_iterator end, int locale_mb_cur_max,
+              const std::string::const_iterator end, size_t locale_mb_cur_max,
               bool locale_utf8locale, std::mbstate_t &state)
 {
   if (locale_mb_cur_max > 1)
